Change KBI0001 title and reframe recommendations
As it became evident in the PR discussion, this KBI is not about
datasets "shared" in terms of write permission, but about datasets
"shared" for reading by users other by directory owners.

Therefore, the scope of the KBI can be described as: operations that
trigger Git's "unsafe repository" mechanism, i.e. DataLad operations
that involve Git working in a directory owned by someone else. This
includes clones made from both repository checkouts and bare
repositories.

Consequently, the recommendation suggests setting the safe.directory
option when cloning from a protected location, but frames it as a
declaration of trust by the user and advices caution.