Add license info to Sphinx conf
This will display the "licensed under CC BY 4.0" text in the footer of
html pages.

It would be great to have CC BY 4.0 be a link to the license deed, but
AFAIK there is no way to add links via conf.py. Other than that, the
footer is generated by the template. We could override the template,
but it seems like too much work.