Add KBI on RIA store improvements in -next
DataLad-next 1.4 patches most of the RIA code, improving it in many
aspects. One particular improvement is related to a `stat` command
error when pushing from macOS to linux. That error has been reported
several times in different places (GitHub issue tracker, Neurostars,
office hour).

This KBI is a rephrasing of my post in a Neurostars thread (1). I
wrote that post mostly to provide a reference to the changes for
future visitors looking for an answer (the response is 2 months older
than the question, asked before the RIA fixes were released). The
answer is now repeated here, to provide a persistent record in a place
controlled by us (in terms of availability) -- granted, Neurostars is
a friendly and trustworthy place (compared to what StackOverflow or
Reddit have become), but it is still an external forum.

(1) https://neurostars.org/t/datalad-push-to-ria-storage-not-working/29049