Add more details about safe.directory config
This adds pointers to Git changelog and commits related to the changes
in Git 2.46 (lead/ing/path/* match) described by @adswa in
edf3f65. The commits have nice messages, and pointing to them fits
well with the rest of the KBI.

A link to the git config docs is updated to point to a section.