Add more information about keyring backends & dependencies
This introduces a small restructuring and extension of KBI0023:
Keyring configuration to address a recently observed case where the
encrypted file keyring was used in the base environment but not a
virtual environment, due to (un)availability of PyCryptodome.

- moved the `datalad wtf --section credentials` example higher; now it
  is part of the introduction, before problems are described

- added an information about listing backends with keyring's CLI

- explained SecretService, EncryptedKeyring, and PlaintextKeyring
  which are likely to come into play

- there are now two problems with separate subheadings; one
  old (SecretService), one new (EncryptedKeyring)

- added a new user solution, explaining that EncryptedKeyring
  requires PyCryptodome, which may not get installed automatically in
  virtualenvs, to become active

Although problems with SecretService and EncryptedKeyring showed up
separately, there is a partial overlap in solutions and a large
overlap in explanations, so it seems more appropriate to extend this
KBI instead of creating a separate one.